requirements.txt
Django==1.8.6
Pillow==3.4.2
pkg-resources==0.0.0
psycopg2==2.6.2
